Use replace instead of add add in FragmentTransaction
When you switch themes, all fragments in the backstack are recreated.
If you use add to go to a new Fragment, the current Fragment's lifecycle methods won't be called properly when you come back to it.
Instead use replace because it will cause the Fragment to be properly re-created once you come back to it.

In this case, the appbar was not being set as actionbar after the user came back to the Overview frgament after switching themes.

Replacing fragments should be fine here since a ViewModel is being used and it retains the data